Finalised Edge Detection algorithm, renamed file, cleaned up code.
Was using edge detected image to find out contours previously, changed to just a blurred thresholded image and the result improved by a lot, also added a max area limit and min area limit to prevent detection of contours that are too big or too small. Cleaned up the code, will use this file as a module and export the final_shape_list list to use in the main program.